% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iashdfs.r
\name{hdfsDelete}
\alias{hdfsDelete}
\title{hdfsDelete function}
\usage{
hdfsDelete(hdfsFilename)
}
\arguments{
\item{hdfsFilename}{Full path to the hdfs file to remove.}
}
\description{
Delete removes the specified file from the supplied HDFS folder
System must have an environment variable defined for the location of the HADOOP command named HADOOP_CMD
}
\examples{
hdfsDelete('/user/tcederquist/mysample.csv') 
}
