% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iashdfs.r
\name{hdfsWebDelete}
\alias{hdfsWebDelete}
\title{hdfsWebDelete function}
\usage{
hdfsWebDelete(host, hdfsFilename, sslVerify = TRUE, verbose = FALSE)
}
\arguments{
\item{host}{Full host and port location for the hdfsFS service. Typically on port 14000, 50470, or 50070}

\item{hdfsFilename}{Full path to the file to read such as /user/myusername/test.csv}

\item{sslVerify}{Optional boolean instruction if the call should expect a valid ssl certificate. TRUE is default, use FALSE for a self signed certificate}
}
\value{
Returns boolean response from hdfsFS
}
\description{
Delete the specified file
}
\examples{
hdfsWebDelete("https://my.server.com:14000", "/user/me/tst2.txt", sslVerify=FALSE)
}
