% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iashdfs.r
\name{hdfsWebLS}
\alias{hdfsWebLS}
\title{hdfsWebLS function}
\usage{
hdfsWebLS(host, path, sslVerify = TRUE)
}
\arguments{
\item{host}{Full host and port location for the hdfsFS service. Typically on port 14000, 50470, or 50070}

\item{path}{Full path to the folder to query such as /user/myusername}

\item{sslVerify}{Optional boolean instruction if the call should expect a valid ssl certificate. TRUE is default, use FALSE for a self signed certificate}
}
\value{
data.frame with the folder information
}
\description{
List the contents of the supplied path and return a data table of files
}
\examples{
listing <- hdfsWebLS('https://my.server.com:14000', '/user/me')
}
