% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iashdfs.r
\name{hdfsWrite}
\alias{hdfsWrite}
\title{hdfsWrite function}
\usage{
hdfsWrite(output, hdfsFilename, append = FALSE)
}
\arguments{
\item{output}{Any string or a function that outputs text directly to the console. This text will be redirected to the hdfs file}

\item{hdfsFilename}{Full path to the hdfs file to write the contents}

\item{append}{Default is FALSE and will delete the file before writing the contents. TRUE will append to the existing file}
}
\description{
Writes the output from the supplied string or function directly to an hdfs file
System must have an environment variable defined for the location of the HADOOP command named HADOOP_CMD
}
\examples{
hdfsWrite(fwrite(mydatatable), "/user/tcederquist/mysample.csv")
hdfsWrite(fwrite(mydatatable[100:200,]), "/user/tcederquist/mysample.csv", append=TRUE)
}
