% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iashdfs.r
\name{ias.hdfsWebBuildURL}
\alias{ias.hdfsWebBuildURL}
\title{ias.hdfsWebBuildURL}
\usage{
ias.hdfsWebBuildURL(host, func, path, op)
}
\arguments{
\item{host}{Host and port of the url 'https://my.server.com:14000}

\item{func}{leading section of the uri '/web/hdfs/v1' for example}

\item{path}{location segment of the hdfsFS request}

\item{op}{Operation parameter to perform on the requested location}
}
\value{
string of the fully formed url
}
\description{
Build a url with basic checks for leading and trailing / marks. Designed for assembling hdfsFS type rest request uris
}
\examples{
ias.hdfsWebBuildURL("https://my.server.com:14000", '/webhdfs/v1', "/user/me", 'LISTSTATUS')
ias.hdfsWebBuildURL("https://my.server.com:14000/", '/webhdfs/v1', "/user/me", 'LISTSTATUS')
}
