% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iashdfs.r
\name{impalaQuickLoad}
\alias{impalaQuickLoad}
\title{impalaQuickLoad function}
\usage{
impalaQuickLoad(srcData, dsn = "", table = "", schema = "",
  hdfsFSHost = "", folder = "/data/tmp", comment = "", sslVerify = TRUE,
  verbose = FALSE)
}
\arguments{
\item{sslVerify}{Optional boolean instruction if the call should expect a valid ssl certificate. TRUE is default, use FALSE for a self signed certificate}

\item{verbose}{Detailed output of the hdfsFS responses}

\item{contents}{String or Steam of data to write into the file}

\item{host}{Full host and port location for the hdfsFS service. Typically on port 14000, 50470, or 50070}

\item{hdfsFilename}{Full path to the file to read such as /user/myusername/test.csv}
}
\value{
Error messages or empty if successful
}
\description{
Write to the contents of the supplied file - creates and overwrites any existing file
}
\examples{
myfile <- hdfsWebRead('https://my.server.com:14000', '/user/me/test.csv')
hdfsWebWrite("Hello there tester2!", "https://my.server.com:14000", "/user/me/test.csv")
}
